requirements.txt
fastapi
uvicorn
python-dotenv
slowapi
python-multipart
